✏️ Outlet으로 레이아웃 분리
