+++ src/App.js
@@ -1,23 +1,20 @@
 import React from "react";
 import { Routes, Route, BrowserRouter } from "react-router-dom";
-import Header from "./pages/Header";
 import Items from "./pages/Items";
 import AddItems from "./pages/AddItems";
+import MainPage from "./pages/MainPage";
 import Main from "./pages/Main";
-import Footer from "./pages/Footer";
 
 function App(props) {
   return (
     <BrowserRouter>
-      <Header />
       <Routes>
-        <Route path="/">
-          <Route index element={<Main />} />
+        <Route path="/" element={<Main />}>
+          <Route index element={<MainPage />} />
           <Route path="items" element={<Items />} />
           <Route path="additems" element={<AddItems />} />
         </Route>
       </Routes>
-      <Footer />
     </BrowserRouter>
   );
 }
